Move container entrypoints to "exec" form
It is the preferred method per the docs here:
https://docs.docker.com/reference/dockerfile#entrypoint

I've also included an `exec` statement inside the scripts.

The need for this - when killing a pod in kubernetes, it sends a kill
signal to the process with pid 1 - which is currently not gizmosql. k8s
then has to wait for the 30 second grace period to end before forcably
killing the container.

Using the `exec` form makes gizmosql be pid 1 inside the container so
k8s can properly shut it down.